// models/Tenant.ts
/**
 * Lagrello API
 * API specification for Lagrello, a simple to use authentication service
 *
 * OpenAPI spec version: 1.0.0
 * Contact: support@lagrello.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { TenantCompanyInfo } from './TenantCompanyInfo';
import { URI } from './URI';
import { HttpFile } from '../http/http';

/**
* Tenant is what you can call your account with lagrello. In your tenant you will store all the information about your account and all your users accounts.
*/
export class Tenant {
    'tenantId': string;
    'tenantName': string;
    'callbackUrl': URI;
    'companyInfo': TenantCompanyInfo;
    'activated': boolean;

    static readonly discriminator: string | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "tenantId",
            "baseName": "tenantId",
            "type": "string",
            "format": "uuid"
        },
        {
            "name": "tenantName",
            "baseName": "tenantName",
            "type": "string",
            "format": ""
        },
        {
            "name": "callbackUrl",
            "baseName": "callbackUrl",
            "type": "URI",
            "format": "uri"
        },
        {
            "name": "companyInfo",
            "baseName": "companyInfo",
            "type": "TenantCompanyInfo",
            "format": ""
        },
        {
            "name": "activated",
            "baseName": "activated",
            "type": "boolean",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return Tenant.attributeTypeMap;
    }
    
    public constructor() {
    }
}